requirements.txt
transformers==4.16.2
tqdm==4.63.0
sklearn==0.0
tensorboard==2.9.0
numpy>=1.22.3
prettytable>=3.0.0
nltk
cotk==0.0.1  # used for glove word vector loading
networkx==3.0
spacy==3.4.4
hydra
faiss-cpu